move OuterCircleTest to ActorTest
